Preprocessor flag for Tombstone
