+++ include/util.h
@@ -479,11 +479,6 @@ static inline int is_occupied(const QF *qf, uint64_t index) {
          1ULL;
 }
 
-static inline int is_tombstone(const QF *qf, uint64_t index) {
-  return (METADATA_WORD(qf, tombstones, index) >>
-          ((index % QF_SLOTS_PER_BLOCK) % 64)) &
-         1ULL;
-}
 
 /* Return num of runends in range [start, start+end). */
 static inline size_t runends_cnt(const QF *qf, size_t start, size_t len) {
@@ -775,19 +770,6 @@ static inline uint64_t find_first_empty_slot(QF *qf, uint64_t from, uint64_t *em
 #endif
 }
 
-/* Find the first tombstone in [from, xnslots), it can be empty or not empty. */
-static inline size_t find_next_tombstone(QF *qf, size_t from) {
-  size_t block_index = from / QF_SLOTS_PER_BLOCK;
-  const size_t slot_offset = from % QF_SLOTS_PER_BLOCK;
-  size_t tomb_offset =
-      bitselectv(get_block(qf, block_index)->tombstones[0], slot_offset, 0);
-  while (tomb_offset == 64) { // No tombstone in the rest of this block.
-    block_index++;
-    tomb_offset = bitselect(get_block(qf, block_index)->tombstones[0], 0);
-  }
-  return block_index * QF_SLOTS_PER_BLOCK + tomb_offset;
-}
-
 /* Return a new word, which first copy `b`, then shift the part (bend, bstart]
  * to the left by `amount`, keep anything out side of the range unchanged.
  * big endian, index from right to left. 
@@ -878,8 +860,10 @@ printf("BL O R T V\n");
            (get_block(qf, i)->occupieds[j / 64] & (1ULL << (j % 64))) ? 1 : 0);
     printf(" %d",
            (get_block(qf, i)->runends[j / 64] & (1ULL << (j % 64))) ? 1 : 0);
+#ifdef QF_TOMBSTONE
     printf(" %d ",
            (get_block(qf, i)->tombstones[j / 64] & (1ULL << (j % 64))) ? 1 : 0);
+#endif
     uint64_t slot = i * QF_SLOTS_PER_BLOCK + j;
     if (slot < qf->metadata->xnslots) {
       printf("%" PRIx64, get_slot(qf, i*QF_SLOTS_PER_BLOCK + j));
@@ -911,10 +895,12 @@ static inline void qf_dump_block(const QF *qf, uint64_t i) {
            (get_block(qf, i)->runends[j / 64] & (1ULL << (j % 64))) ? 1 : 0);
   printf("\n");
 
+#ifdef QF_TOMBSTONE
   for (j = 0; j < QF_SLOTS_PER_BLOCK; j++)
     printf(" %d ",
            (get_block(qf, i)->tombstones[j / 64] & (1ULL << (j % 64))) ? 1 : 0);
   printf("\n");
+#endif
 
 #if QF_BITS_PER_SLOT == 8 || QF_BITS_PER_SLOT == 16 || QF_BITS_PER_SLOT == 32
   for (j = 0; j < QF_SLOTS_PER_BLOCK; j++)
@@ -968,43 +954,6 @@ static inline void shift_runends(QF *qf, int64_t first, uint64_t last,
       0, METADATA_WORD(qf, runends, 64 * last_word), bstart, bend, distance);
 }
 
-/* Shift metadata runends and tombstones in range [first, last) to the big
- * direction by distance.
- * `last` to `last+distance-1` will be replaced. Fill with 0s in the small size.
- */
-static inline void shift_runends_tombstones(QF *qf, int64_t first,
-                                            uint64_t last, uint64_t distance) {
-  assert(last < qf->metadata->xnslots);
-  assert(distance < 64);
-  uint64_t first_word = first / 64;
-  uint64_t bstart = first % 64;
-  uint64_t last_word = (last + distance - 1) / 64;
-  uint64_t bend = (last + distance - 1) % 64 + 1;
-
-  if (last_word != first_word) {
-    METADATA_WORD(qf, runends, 64 * last_word) = shift_into_b(
-        METADATA_WORD(qf, runends, 64 * (last_word - 1)),
-        METADATA_WORD(qf, runends, 64 * last_word), 0, bend, distance);
-    METADATA_WORD(qf, tombstones, 64 * last_word) = shift_into_b(
-        METADATA_WORD(qf, tombstones, 64 * (last_word - 1)),
-        METADATA_WORD(qf, tombstones, 64 * last_word), 0, bend, distance);
-    bend = 64;
-    last_word--;
-    while (last_word != first_word) {
-      METADATA_WORD(qf, runends, 64 * last_word) = shift_into_b(
-          METADATA_WORD(qf, runends, 64 * (last_word - 1)),
-          METADATA_WORD(qf, runends, 64 * last_word), 0, bend, distance);
-      METADATA_WORD(qf, tombstones, 64 * last_word) = shift_into_b(
-          METADATA_WORD(qf, tombstones, 64 * (last_word - 1)),
-          METADATA_WORD(qf, tombstones, 64 * last_word), 0, bend, distance);
-      last_word--;
-    }
-  }
-  METADATA_WORD(qf, runends, 64 * last_word) = shift_into_b(
-      0, METADATA_WORD(qf, runends, 64 * last_word), bstart, bend, distance);
-  METADATA_WORD(qf, tombstones, 64 * last_word) = shift_into_b(
-      0, METADATA_WORD(qf, tombstones, 64 * last_word), bstart, bend, distance);
-}
 
 static inline int
 remove_replace_slots_and_shift_remainders_and_runends_and_offsets(
@@ -1138,32 +1087,6 @@ static inline void quotien_remainder(const QF *qf, const uint64_t hash,
  * run_end_index 
  *    0: if didn't find it, the index and range would be where to insert it.
  */
-static int find(const QF *qf, const uint64_t quotient, const uint64_t remainder,
-                uint64_t *const index, uint64_t *const run_start_index,
-                uint64_t *const run_end_index) {
-  *run_start_index = 0;
-  if (quotient != 0)
-    *run_start_index = run_end(qf, quotient - 1) + 1;
-  *index = *run_start_index;
-  if (!is_occupied(qf, quotient)) {
-    // no such run
-    *run_end_index = *run_start_index;
-    return 0;
-  }
-  *run_end_index = run_end(qf, quotient) + 1;
-  uint64_t curr_remainder;
-  do {
-    if (!is_tombstone(qf, *index)) {
-      curr_remainder = get_slot(qf, *index) >> qf->metadata->value_bits;
-      if (remainder == curr_remainder)
-        return 1;
-      if (remainder < curr_remainder)
-        return 0;
-    }
-    *index += 1;
-  } while (*index < *run_end_index);
-  return 0;
-}
 
 static inline int64_t bitscanforward(uint64_t val)
 {
